requirements.txt
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
dicttoxml==1.7.4
idna==3.2
mysql-connector-python==8.0.26
mysqlclient==2.0.3
protobuf==3.17.3
pycodestyle==2.7.0
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
xmltodict==0.12.0
